+++ app/controllers/shift_requests_controller.rb
@@ -4,6 +4,7 @@ class ShiftRequestsController < ApplicationController
     if staff_user_signed_in?
       @shift_requests = current_staff_user.shift_requests
       @shifts = current_staff_user.shifts
+      @shift_request = ShiftRequest.new
     end
   end
 end
