+++ app/javascript/packs/application.js
@@ -6,11 +6,10 @@
 require("@rails/ujs").start()
 require("@rails/activestorage").start()
 require("channels")
+require("../staff_calendar")
 require("../switch")
 require("../modal_board")
 require("../input_display")
-
-
 
 // Uncomment to copy all static images under ../images to the output folder and reference
 // them with the image_pack_tag helper in views (e.g <%= image_pack_tag 'rails.png' %>)
