+++ app/javascript/switch.js
@@ -22,31 +22,3 @@ if (typeof gon !== 'undefined') {
     });
   }
 }
-if (!location.pathname.match("/boards")) {
-  document.addEventListener("DOMContentLoaded", () => {
-    const requestCalendar = document.getElementById("requested-shift-calendar");
-    const fixedCalendar = document.getElementById("fixed-shift-calendar");
-    const requestBtnF = document.getElementById("f_r-btn");
-    const fixedBtnR = document.getElementById("r_f-btn");
-    const requestBtnR = document.getElementById("r_r-btn");
-    const fixedBtnF = document.getElementById("f_f-btn");
-    fixedBtnF.style.backgroundColor = "#00a1e9";
-    fixedBtnF.style.color = "#ffffff";
-
-    requestBtnF.addEventListener('click', () => {
-      requestCalendar.style.display = "block";
-      fixedCalendar.style.display = "none";
-      fixedBtnR.removeAttribute("style");
-      requestBtnR.style.backgroundColor = "#00a1e9";
-      requestBtnR.style.color = "#ffffff";
-    });
-    fixedBtnR.addEventListener('click', () => {
-      console.log(2);
-      requestCalendar.style.display = "none";
-      fixedCalendar.style.display = "block";
-      requestBtnF.removeAttribute("style");
-      fixedBtnF.style.backgroundColor = "#00a1e9";
-      fixedBtnF.style.color = "#ffffff";
-    });
-  });
-}
